big rewrite of the keyboard toolbar handling
The idea was to have smoother animation of the keyboard toolbar. To do that i had to directly call the animation in the KeyboardWillShow without delay. But to do that i had to make sure the accessory view was set, so i had to modify things so that the proxy getting the focus call didGetFocus even before actually getting the focus.